Fixed Mesh Loading
For some reason my textures are all flipped. inverting the Y component
of the Texture coordinate works. Might have to do with how the image is
loaded, but I'm not sure how.

Also added a fix for files that have verticies and normals but no
textures, since I ran into that in one test.