Removed Duplicate Code from Text
As most of the functions I was using were duplicates of those in
BaseGLObject, they were removed and replaced with calls to the internal
methods. Within BaseGLObject, the index count was moved to the
InitializeVertexObject Method, since derived classes won't always run
the constructor and that information will be present in this method
regardless. Also added check for the Model Matrix Uniform update. Not
that it really matters, but for now it seems like a good idea not to
bother.