Somewhat fixed raycasts from the mouse
Created a sub class within BaseGLObject to hold data about the collision
model. At the moment, unless it is specifically loaded, it just uses the
vertex data from the model that was loaded. Provides a CheckForCollision
method that accepts a raycast, and determines if the ray will intersect
any of the triangles.

Raycasting mostly works, but there seems to be some issue regarding the
y axis. To be looked into.